+++ gpype/backend/sources/bci_core8.py
@@ -26,7 +26,7 @@ class BCICore8(AmplifierSource):
     """
 
     #: Source code fingerprint for licensing verification
-    FINGERPRINT = "5d30ffdac8e9f29f4e7c0a690bfba9ce"
+    FINGERPRINT = "a08e74f489d777c589e562d2ddff0d42"
 
     #: Optional buffer level monitoring output port name
     PORT_BUF_LEVEL = "buffer_level"